Big Improvement in MainNavHost with FeedDrawer

+++ app/src/main/java/com/example/scrollbooker/screens/inbox/InboxScreen.kt
@@ -16,7 +16,7 @@ import com.example.scrollbooker.screens.inbox.components.NotificationsList
 @Composable
 fun InboxScreen(
     viewModel: InboxViewModel,
-    onNavigate: () -> Unit
+    onNavigate: (Int) -> Unit
 ) {
     val notifications = viewModel.notifications.collectAsLazyPagingItems()
     val refreshState = notifications.loadState.refresh
